% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_description.R
\name{description_encoding}
\alias{description_encoding}
\title{Obtain Encoding from DESCRIPTION file}
\usage{
description_encoding(package)
}
\arguments{
\item{package}{Data frame with package metadata or string with package name.}
}
\value{
Package Encoding.
}
\description{
Obtain Encoding from DESCRIPTION file
}
\examples{
description_encoding("tools")
description_encoding("dplyr")
}
