% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_description.R
\name{.description_element}
\alias{.description_element}
\title{Get element from DESCRIPTION file}
\usage{
.description_element(description, label)
}
\arguments{
\item{description}{DESCRIPTION file contents.}

\item{label}{String with label in the DESCRIPTION content.}
}
\value{
Value from the DESCRIPTION file.
}
\description{
Get element from DESCRIPTION file
}
\keyword{internal}
