% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cran.R
\name{.list_notes.character}
\alias{.list_notes.character}
\title{List notes for a given package for a given path}
\usage{
.list_notes.character(package)
}
\arguments{
\item{package}{String with path to package.}
}
\value{
Data frame with metadata for each note, including: Rd file name,
    line number, category (e.g., Lost braces') and note details.
}
\description{
List notes for a given package for a given path
}
\keyword{internal}
