% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cran.R
\name{.list_notes.data.frame}
\alias{.list_notes.data.frame}
\title{List notes for a given package for a given package metadata data frame}
\usage{
.list_notes.data.frame(package)
}
\arguments{
\item{package}{Data frame with package metadata.}
}
\value{
Data frame with metadata for each note, including: Rd file name,
    line number, category (e.g., Lost braces') and note details.
}
\description{
List notes for a given package for a given package metadata data frame
}
\keyword{internal}
