% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cran.R
\name{list_notes}
\alias{list_notes}
\title{List notes for a given package}
\usage{
list_notes(package)
}
\arguments{
\item{package}{Either a row from `list_packages_with_notes()` or a path to a
package (notes will be generated with `tools::checkRd()`).}
}
\value{
Data frame with metadata for each note, including: Rd file name,
    line number, category (e.g., Lost braces') and note details.
}
\description{
List notes for a given package
}
\examples{
list_packages_with_notes() |>
  dplyr::slice(1) |>
  list_notes()
}
