% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cran.R
\name{list_packages_with_notes}
\alias{list_packages_with_notes}
\title{List packages on CRAN with documentation notes}
\usage{
list_packages_with_notes()
}
\value{
Data frame with list of packages and associated metadata
}
\description{
List packages on CRAN with documentation notes
}
